Split bsmfilter.h and bsmfilter.cpp into 8 files.
