+++ src/bsmHandler.cpp
@@ -38,254 +38,6 @@
 #include "spdlog/spdlog.h"
 #include "./redaction-properties/RedactionPropertiesManager.cpp"
 
-/** Start IdRedactor */
-
-IdRedactor::IdRedactor() :
-    inclusion_set_{},
-    redacted_value_{"FFFFFFFF"},                    // default value.
-    inclusions_{false}                              // redact everything.
-{
-    // setup random number generator.
-    std::random_device rd;
-    rgen_ = std::mt19937{ rd() };
-    dist_ = std::uniform_int_distribution<uint32_t>{ 0, std::numeric_limits<uint32_t>::max() };
-}
-
-IdRedactor::IdRedactor( const ConfigMap& conf ) :
-    IdRedactor{}
-{
-    // TODO: The redaction value is deprecated (it is randomly assigned now).
-    auto search = conf.find("privacy.redaction.id.value");
-    if ( search != conf.end() ) {
-        redacted_value_ = search->second;
-    }
-
-    search = conf.find("privacy.redaction.id.inclusions");
-    if ( search != conf.end() && search->second=="ON" ) {
-        inclusions_ = true;
-    }
-
-    search = conf.find("privacy.redaction.id.included");
-    if ( search != conf.end() ) {
-        StrVector sv = string_utilities::split( search->second, ',' );
-        for ( auto& id : sv ) {
-            inclusion_set_.insert( id );
-        }
-    }
-};
-
-bool IdRedactor::HasInclusions() const
-{
-    return inclusions_;
-}
-
-int IdRedactor::NumInclusions() const
-{
-    if ( inclusions_ ) {
-        return inclusion_set_.size();
-    }
-    return -1;
-}
-
-void IdRedactor::RedactAll()
-{
-    inclusion_set_.clear();
-    inclusions_ = false;
-}
-
-bool IdRedactor::ClearInclusions()
-{
-    bool r = inclusion_set_.size() > 0;
-    inclusion_set_.clear();
-    return r;
-}
-
-bool IdRedactor::AddIdInclusion( const std::string& id )
-{
-    auto result = inclusion_set_.insert( id );
-    if ( !inclusions_ && result.second ) {
-        // previously redacting everything, not we are building the inclusion list.
-        inclusions_ = true;
-    }
-    return result.second;
-}
-
-bool IdRedactor::RemoveIdInclusion( const std::string& id )
-{
-    bool r = false;
-    auto search = inclusion_set_.find( id );
-    if ( search != inclusion_set_.end() ) {
-        // id is currently in the inclusion set so erase it.
-        inclusion_set_.erase( search );
-        r = true;
-    }
-    return r;
-}
-
-std::string IdRedactor::GetRandomId()
-{
-	std::stringstream ss;
-    uint32_t v = dist_(rgen_);
-    ss << std::hex << std::setfill('0') << std::setw(sizeof(uint32_t)*2) << v;                                                 
-	return ss.str();
-}
-
-bool IdRedactor::operator()( std::string& id )
-{
-    if ( inclusions_ ) {
-        auto search = inclusion_set_.find( id );
-        if ( search == inclusion_set_.end() ) {
-            // Case 1: Using inclusion set, but not found; do NOT redact.
-            return false;
-        }
-        // Case 2: Found this id in the inclusions set; redact.
-    } // Case 3: Not using inclusion set; redact everything.
-
-    // Case 2 and 3: Overwrite existing id with redaction id.
-    //id = redacted_value_;
-    id = GetRandomId();
-    return true;
-}
-
-const std::string& IdRedactor::redaction_value() const
-{
-    return redacted_value_;
-}
-
-/** Start VelocityFilter */
-
-VelocityFilter::VelocityFilter() :
-    min_{kDefaultMinVelocity},
-    max_{kDefaultMaxVelocity}
-{}
-
-VelocityFilter::VelocityFilter( const ConfigMap& conf ) :
-    VelocityFilter{}
-{
-    auto search = conf.find("privacy.filter.velocity.min");
-    if ( search != conf.end() ) {
-        min_ = std::stod( search->second );
-    }
-
-    search = conf.find("privacy.filter.velocity.max");
-    if ( search != conf.end() ) {
-        max_ = std::stod( search->second );
-    }
-}
-
-void VelocityFilter::set_min( double v ) {
-    min_ = v;
-}
-
-void VelocityFilter::set_max( double v ) {
-    max_ = v;
-}
-
-bool VelocityFilter::operator()( double v ) {
-    // true = filter it!
-    return v < min_ || v > max_;
-}
-
-bool VelocityFilter::suppress( double v ) {
-    return (*this)(v);
-}
-
-bool VelocityFilter::retain( double v ) {
-    return !(*this)(v);
-}
-
-/** Start BSM */
-
-BSM::BSM() :
-    geo::Point{90.0, 180.0},
-    velocity_{ -1 },
-    dsec_{0},
-    id_{""},
-    oid_{""},
-    partII_{""},
-    logstring_{}
-{}
-
-void BSM::reset() {
-    lat = 90.0;
-    lon = 180.0;
-    velocity_ = -1.0;
-    dsec_ = 0;
-    id_ = "";
-    oid_ = "";
-    partII_ = "";
-}
-
-std::string BSM::logString() {
-    logstring_ = "(" + id_;
-    logstring_ += "," + std::to_string(dsec_);
-    logstring_ += "," + std::to_string(lat);
-    logstring_ += "," + std::to_string(lon);
-    logstring_ += "," + std::to_string(velocity_);
-    logstring_ += ")";
-    return logstring_;
-}
-
-void BSM::set_velocity( double v ) {
-    velocity_ = v;
-}
-
-double BSM::get_velocity() const {
-    return velocity_;
-}
-
-void BSM::set_latitude( double latitude ) {
-    lat = latitude;
-}
-
-void BSM::set_longitude( double longitude ) {
-    lon = longitude;
-}
-
-void BSM::set_secmark( uint16_t dsec ) {
-    dsec_ = dsec;
-}
-
-uint16_t BSM::get_secmark() const {
-    return dsec_;
-}
-
-void BSM::set_id( const std::string& s ) {
-    id_ = s;
-}
-
-const std::string& BSM::get_id() const {
-    return id_;
-}
-
-// for testing.
-void BSM::set_original_id( const std::string& s ) {
-    oid_ = s;
-}
-
-// for testing.
-const std::string& BSM::get_original_id() const {
-    return oid_;
-}
-
-void BSM::set_partII( const std::string& s ) {
-    partII_ = s;
-}
-
-const std::string& BSM::get_partII() const {
-    return partII_;
-}
-
-std::ostream& operator<<( std::ostream& os, const BSM& bsm )
-{
-    os  << std::setprecision(16) << "Pos: (" << bsm.lat << ", " << bsm.lon << "), ";
-    os  << "Spd: "  << bsm.velocity_ << " ";
-    os  << "Id: "  <<  bsm.id_;
-    return os;
-}
-
-/** Start BSMHandler */
-
 BSMHandler::ResultStringMap BSMHandler::result_string_map{
             { ResultStatus::SUCCESS, "success" },
             { ResultStatus::SPEED, "speed" },
